Update how-to-harden-ubuntu-server.md
After generating SSH keys and logging in to test the new configuration, the guide says to test the configuration using key.pub -- I think this is a mistake because that's the public part of the key.  Using <ssh key>.pub results in:
@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
@         WARNING: UNPROTECTED PRIVATE KEY FILE!          @
@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
Permissions 0644 for <ssh key> are too open.
It is required that your private key files are NOT accessible by others.
This private key will be ignored.

Using the sshkey without pub works just fine.