Note for 18089
I don't like this solution.  A better one is to update the graphic to show port 18081/18089, and add a section about how one ought to choose between the two, or, perhaps, use the default (18081) for unrestricted local access (127.0.0.1) and 18089 for restricted public access.  In fact, given the default use of 18081 for unrestricted rpc access, I think the graphic is dangerous.