Update finding UTXOs to ignore native tokens
The current loop that retrieves all the available ADA is using every UTXO. Some UTXOs are native tokens (NFTs) and should not be considered in the tx_in. If they are used in the tx_in they also need to be added to the tx_out, causing unnecessary sending/receiving, also complicating the script wildly. If this solution is not acceptable, then we could consider not using a loop, and simply taking the first UTXO, then later checking if the balance is enough after adding the fee.