Update hardening-an-ubuntu-server.md
Hello. I've added a ::fire:: note to clarify that the proposed iptables rules are forgotten upon reboot. 
Furthermore, I have added a reference to a script which can enhance the security level and protect against a wide range of DDoS attacks. This script has been tested/developed during a DDoS attack and made available to the community.